Initial Commit
This app functions on the principle that the user has ingredients to
make drinks and can then use the app to look up what drinks they can
make. The user is presented with a list of ingredients, divided by
category, where they select those that they have on hand. After tapping
the search button they are presented with another list of drinks that
can be made with the ingredients they have selected. A drink may be
selected to view more detailed information on ingredients, instructions,
descriptions, etc. Drinks can be favorited from this view, and a list of
favorite drinks can be found from the ingredient selection view. There
are tabs along the bottom that allow for the addition of more drinks and
ingredients if the user was unable to find them.